+++ helper/logger.py
@@ -5,14 +5,22 @@ import os.path
 
 class Logger():
     def __init__(self, name):
-        FORMAT = '%(asctime)-15s %(message)s'
-        formatter = logging.Formatter(FORMAT)
         self.log_adapter = logging.getLogger(name)
-        self.log_adapter.setLevel(logging.DEBUG)
-        fh = logging.FileHandler(
+        self.set_level()
+        format = '%(asctime)-15s %(message)s'
+        formatter = logging.Formatter(format)
+        self.file_handler = logging.FileHandler(
             './logger/logger-{}.log'.format(datetime.now().strftime('%Y-%m-%d')), 'a')
-        fh.setFormatter(formatter)
-        self.log_adapter.addHandler(fh)
+        self.file_handler.setFormatter(formatter)
+        self.log_adapter.addHandler(self.file_handler)
+
+    def set_level(self, level='d'):
+        if level == 'd':
+            self.log_adapter.setLevel(logging.DEBUG)
+        elif level == 'i':
+            self.log_adapter.setLevel(logging.INFO)
+        elif level == 'e':
+            self.log_adapter.setLevel(logging.ERROR)
 
     def d(self, msg):
         self.log_adapter.debug(msg)
@@ -20,14 +28,8 @@ class Logger():
     def i(self, msg):
         self.log_adapter.info(msg)
 
-    def w(self, msg):
-        self.log_adapter.warning(msg)
-
     def e(self, msg):
         self.log_adapter.error(msg)
-
-    def c(self, msg):
-        self.log_adapter.critical(msg)
 
 
 log = Logger('main')